[GH-20] Move logic into libs
